Simplify the behavior change of the bootstrap account.
